read double from txt file

+++ mainwindow.cpp
@@ -100,7 +100,7 @@ void MainWindow::on_action_Open_triggered()
 
     while(!txtStream.atEnd()){
         txtStream >> tmp;
-        data_input->data[i][j] = QString(tmp).toInt();//+(double)(rand()%100)/100;
+        data_input->data[i][j] = QString(tmp).toDouble();//+(double)(rand()%100)/100;
         //data_input->data[i][j] = cos(2*M_PI*(0.12*i+0.13*j));
         //if(cos(2*M_PI*(0.143*i+0.143*j))>0) data_input->data[i][j] = 1;
         //if(cos(2*M_PI*(0.143*i+0.143*j))<=0) data_input->data[i][j] = 0;
@@ -142,7 +142,7 @@ void MainWindow::on_action_openFFT_triggered()
 
     while(!txtStream.atEnd()){
         txtStream >> tmp;
-        data_fft->data[i][j] = QString(tmp).toInt();
+        data_fft->data[i][j] = QString(tmp).toDouble();
         i++;
         if(i>=data_fft->size_x){
             i=0; j++;
